+++ src/PetControlSystem.Domain/Services/CustomerService.cs
@@ -17,25 +17,69 @@ namespace PetControlSystem.Domain.Services
         {
             if (ExecuteValidation(new CustomerValidation(), customer)) return;
 
+            if (_customerRepository.Get(c => c.Document == customer.Document).Result.Any())
+            {
+                Notify("There is already a customer with this document");
+                return;
+            }
+
+            if (_customerRepository.Get(c => c.Document == customer.Email).Result.Any())
+            {
+                Notify("There is already a customer with this email");
+                return;
+            }
+
             await _customerRepository.Add(customer);
         }
 
         public async Task Delete(Guid id)
         {
+            if (_customerRepository.GetById(id) is null)
+            {
+                Notify("There is no customer with this ID");
+                return;
+            }
+
             await _customerRepository.Remove(id);
         }
 
         public async Task Update(Customer customer)
         {
             if (ExecuteValidation(new CustomerValidation(), customer)) return;
+            
+            if (_customerRepository.GetById(customer.Id) is null)
+            {
+                Notify("There is no customer with this ID");
+                return;
+            }
 
             await _customerRepository.Update(customer);
         }
 
-        public async Task<Customer> GetById(Guid id)
+        public async Task<Customer?> GetById(Guid id)
         {
-            await _customerRepository.GetById(id);
-            return null;
+            var result = await _customerRepository.GetById(id);
+
+            if (result is null)
+            {
+                Notify("There is no customer with this ID");
+                return null;
+            }
+;
+            return result;
+        }
+
+        public async Task<IEnumerable<Customer>?> GetAll()
+        {
+            var result = await _customerRepository.GetAll();
+
+            if (result is null)
+            {
+                Notify("There are no customers");
+                return null;
+            }
+
+            return result;
         }
 
         public void Dispose()
